Update sln and move to right place
