+++ sources/Adapters/SDL/GUI/SDLGUIWindowImp.cpp
@@ -2,7 +2,7 @@
 #include "UIFramework/SimpleBaseClasses/GUIWindow.h"
 #include "System/System/System.h"
 #include "Application/Model/Config.h"
-#include "Application/Model/FontConfig.h"
+#include "UIFramework/BasicDatas/FontConfig.h"
 #include "System/Console/Trace.h"
 #include <string.h>
 #include "System/Console/n_assert.h"
